--- a/dev-app.config.js
+++ b/dev-app.config.js
@@ -1,0 +1,6 @@
+const merge = require('webpack-merge');
+const devConfig = require('./dev.config.js');
+
+module.exports = merge.strategy({ entry: 'replace' })(devConfig, {
+  entry: { renderer: './app/app.ts' },
+});