--- a/package.json
+++ b/package.json
@@ -217,6 +217,7 @@
     "webdriverio": "4.14.1",
     "webpack": "4.26.1",
     "webpack-cli": "3.1.2",
+    "webpack-merge": "^4.2.1",
     "which": "1.3.1"
   },
   "resolutions": {