--- a/package.json
+++ b/package.json
@@ -11,7 +11,7 @@
   },
   "scripts": {
     "compile": "yarn clear && yarn webpack-cli --progress --config dev.config.js",
-    "compile:ci": "yarn clear && yarn webpack-cli --config dev.config.js",
+    "compile:ci": "yarn clear && yarn webpack-cli --config ci.config.js",
     "compile:production": "yarn clear && yarn webpack-cli --progress --config prod.config.js",
     "webpack-cli": "node --max-old-space-size=4096 node_modules/webpack-cli/bin/cli.js",
     "watch": "yarn clear && yarn webpack-cli --watch --progress --config dev.config.js",