@andycreeth 2019-04-10
Hmm, I think we either want the production one to extend the dev one, or we need a base config that they both extend.  Otherwise we have far too much duplication between the two and they will get out of sync.