@gettinToasty 2019-04-10
@andycreeth that should be totally do-able, lemme see if i can get to a good solution there 👍